+++ core/camel-core/src/main/java/org/apache/camel/impl/AbstractCamelContext.java
@@ -98,6 +98,7 @@ import org.apache.camel.model.rest.RestDefinition;
 import org.apache.camel.model.rest.RestsDefinition;
 import org.apache.camel.model.transformer.TransformerDefinition;
 import org.apache.camel.model.validator.ValidatorDefinition;
+import org.apache.camel.processor.DefaultDeferServiceFactory;
 import org.apache.camel.processor.MulticastProcessor;
 import org.apache.camel.processor.interceptor.Debug;
 import org.apache.camel.processor.interceptor.HandleFault;
@@ -113,6 +114,7 @@ import org.apache.camel.spi.DataFormat;
 import org.apache.camel.spi.DataFormatResolver;
 import org.apache.camel.spi.DataType;
 import org.apache.camel.spi.Debugger;
+import org.apache.camel.spi.DeferServiceFactory;
 import org.apache.camel.spi.EndpointRegistry;
 import org.apache.camel.spi.EndpointStrategy;
 import org.apache.camel.spi.EventNotifier;
@@ -272,6 +274,7 @@ public abstract class AbstractCamelContext extends ServiceSupport implements Mod
     private volatile ReloadStrategy reloadStrategy;
     private volatile RouteController routeController;
     private volatile ScheduledExecutorService errorHandlerExecutorService;
+    private final DeferServiceFactory deferServiceFactory = new DefaultDeferServiceFactory();
 
     private TransformerRegistry<TransformerKey> transformerRegistry;
     private ValidatorRegistry<ValidatorKey> validatorRegistry;
@@ -3920,6 +3923,11 @@ public abstract class AbstractCamelContext extends ServiceSupport implements Mod
     @Override
     public void setHeadersMapFactory(HeadersMapFactory headersMapFactory) {
         this.headersMapFactory = doAddService(headersMapFactory);
+    }
+
+    @Override
+    public DeferServiceFactory getDeferServiceFactory() {
+        return deferServiceFactory;
     }
 
     protected Map<String, RouteService> getRouteServices() {
